requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.9
django-crispy-forms==1.6.0
django-filter==1.0.4
django-js-reverse==0.7.3
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-storages==1.6.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
heroku==0.1.4
idna==2.5
Markdown==2.6.8
markdown2==2.3.1
Pillow==3.1.0
PyJWT==1.5.2
python-dateutil==1.5
requests==2.18.1
urllib3==1.21.1
whitenoise==3.3.0